Activity 명 변경, 추가
