+++ app/src/main/java/com/example/muhanparking/LoginActivity.java
@@ -1,0 +1,68 @@
+package com.example.muhanparking;
+
+import android.os.Bundle;
+import android.content.Intent;
+import androidx.activity.EdgeToEdge;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.core.graphics.Insets;
+import androidx.core.view.ViewCompat;
+import androidx.core.view.WindowInsetsCompat;
+
+import android.text.TextUtils;
+import android.view.View;
+import android.widget.Button;
+import android.widget.EditText;
+import android.widget.Toast;
+
+public class LoginActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        EdgeToEdge.enable(this);
+        setContentView(R.layout.activity_main);
+        ViewCompat.setOnApplyWindowInsetsListener(findViewById(R.id.main), (v, insets) -> {
+            Insets systemBars = insets.getInsets(WindowInsetsCompat.Type.systemBars());
+            v.setPadding(systemBars.left, systemBars.top, systemBars.right, systemBars.bottom);
+            return insets;
+        });
+
+        EditText etId = findViewById(R.id.etId);
+        EditText etPassword = findViewById(R.id.etPassword);
+        Button btnLogin = findViewById(R.id.btnLogin);
+        Button btnRegister = findViewById(R.id.btnRegister);
+
+        btnLogin.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                String email = etId.getText().toString();
+                String password = etPassword.getText().toString();
+
+                // 입력값 검증
+                if (TextUtils.isEmpty(email) || TextUtils.isEmpty(password)) {
+                    Toast.makeText(LoginActivity.this, "Please fill in all fields.", Toast.LENGTH_SHORT).show();
+                } else {
+                    // 간단한 로그인 검증 (하드코딩된 값과 비교)
+                    String validId = "test";
+                    String validPassword = "1234";
+
+                    if (email.equals(validId) && password.equals(validPassword)) {
+                        Intent intent = new Intent(LoginActivity.this, MainActivity.class);
+                        startActivity(intent);  // 새 Activity 시작
+                        finish();  // 현재 Activity 종료 (선택 사항)
+                    } else {
+                        Toast.makeText(LoginActivity.this, "Invalid Email or Password.", Toast.LENGTH_SHORT).show();
+                    }
+                }
+            }
+        });
+
+        btnRegister.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                Intent intent = new Intent(LoginActivity.this, Register_Activity.class);
+                startActivity(intent);
+            }
+        });
+    }
+}
