+++ app/src/main/java/com/example/muhanparking/MainActivity.java
@@ -1,18 +1,16 @@
 package com.example.muhanparking;
 
 import android.os.Bundle;
-import android.content.Intent;
+import android.view.View;
+import android.widget.ImageView;
+import android.widget.Button;
 import androidx.activity.EdgeToEdge;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.core.graphics.Insets;
 import androidx.core.view.ViewCompat;
 import androidx.core.view.WindowInsetsCompat;
-import android.os.Bundle;
-import android.text.TextUtils;
-import android.view.View;
-import android.widget.Button;
-import android.widget.EditText;
-import android.widget.Toast;
+
+import android.content.Intent;
 
 public class MainActivity extends AppCompatActivity {
 
@@ -20,48 +18,86 @@ public class MainActivity extends AppCompatActivity {
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         EdgeToEdge.enable(this);
-        setContentView(R.layout.activity_main);
+        setContentView(R.layout.activity_main2);
         ViewCompat.setOnApplyWindowInsetsListener(findViewById(R.id.main), (v, insets) -> {
             Insets systemBars = insets.getInsets(WindowInsetsCompat.Type.systemBars());
             v.setPadding(systemBars.left, systemBars.top, systemBars.right, systemBars.bottom);
             return insets;
         });
 
-        EditText etId = findViewById(R.id.etId);
-        EditText etPassword = findViewById(R.id.etPassword);
-        Button btnLogin = findViewById(R.id.btnLogin);
-        Button btnRegister = findViewById(R.id.btnRegister);
+        ImageView myInfo = findViewById(R.id.btn_myinfo);
+        ImageView parking = findViewById(R.id.btn_parking);
+        ImageView regular = findViewById(R.id.btn_regular_parking);
+        ImageView menuIcon = findViewById(R.id.menu_icon);
+        ImageView settingsIcon = findViewById(R.id.settings_icon);
+        ImageView notiIcon = findViewById(R.id.noti_icon);
+        Button more = findViewById(R.id.btn_more);
 
-        btnLogin.setOnClickListener(new View.OnClickListener() {
+        menuIcon.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                String email = etId.getText().toString();
-                String password = etPassword.getText().toString();
-
-                // 입력값 검증
-                if (TextUtils.isEmpty(email) || TextUtils.isEmpty(password)) {
-                    Toast.makeText(MainActivity.this, "Please fill in all fields.", Toast.LENGTH_SHORT).show();
-                } else {
-                    // 간단한 로그인 검증 (하드코딩된 값과 비교)
-                    String validId = "test";
-                    String validPassword = "1234";
-
-                    if (email.equals(validId) && password.equals(validPassword)) {
-                        Intent intent = new Intent(MainActivity.this, MainActivity2.class);
-                        startActivity(intent);  // 새 Activity 시작
-                        finish();  // 현재 Activity 종료 (선택 사항)
-                    } else {
-                        Toast.makeText(MainActivity.this, "Invalid Email or Password.", Toast.LENGTH_SHORT).show();
-                    }
-                }
+                // 햄버거 메뉴 클릭 시 실행될 동작
+                Intent intent = new Intent(MainActivity.this, Menu_Activity.class);
+                startActivity(intent);
+                finish();
             }
         });
 
-        btnRegister.setOnClickListener(new View.OnClickListener() {
+        notiIcon.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                Intent intent = new Intent(MainActivity.this, Notifications_Activity.class);
+                startActivity(intent);
+                finish();
+
+            }
+        });
+
+        parking.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                Intent intent = new Intent(MainActivity.this, Simulation_Activity.class);
+                startActivity(intent);
+                finish();
+            }
+        });
+
+
+        settingsIcon.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                Intent intent = new Intent(MainActivity.this, Register_Activity.class);
+                // 설정 버튼 클릭 시 실행될 동작
+                Intent intent = new Intent(MainActivity.this, Setting_Activity.class);
                 startActivity(intent);
+                finish();
+            }
+        });
+
+        regular.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                Intent intent = new Intent(MainActivity.this, Regular_Parking_Activity.class);
+                startActivity(intent);
+                finish();
+            }
+        });
+
+        myInfo.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                Intent intent = new Intent(MainActivity.this, Mypage_Activity.class);
+                startActivity(intent);
+                finish();
+            }
+        });
+
+        //자세히 보기 버튼
+        more.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                Intent intent = new Intent(MainActivity.this, Empty_Parking_Activity.class);
+                startActivity(intent);
+                finish();
             }
         });
     }
