+++ app/src/main/java/com/example/muhanparking/Simulation_Activity.java
@@ -29,7 +29,7 @@ public class Simulation_Activity extends AppCompatActivity {
         back.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                Intent intent = new Intent(Simulation_Activity.this, MainActivity2.class);
+                Intent intent = new Intent(Simulation_Activity.this, MainActivity.class);
                 startActivity(intent);
                 finish();
             }
