+++ app/src/main/res/layout/activity_main2.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainActivity2">
+    tools:context=".MainActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
